Roll out CTest-based testing
🧙 Conjured by AI via [pi.dev](https://pi.dev/) using gpt-5.5
